+++ packages/preferences-persistence/src/migrations/preferences-package-data/convert-editor-settings.js
@@ -1,0 +1,41 @@
+/**
+ * Internal dependencies
+ */
+
+export default function convertEditorSettings( data ) {
+	let newData = data;
+	const settingsToMoveToCore = [
+		'allowRightClickOverrides',
+		'keepCaretInsideBlock',
+		'showBlockBreadcrumbs',
+		'showIconLabels',
+		'showListViewByDefault',
+	];
+
+	settingsToMoveToCore.forEach( ( setting ) => {
+		if ( data?.[ 'core/edit-post' ]?.[ setting ] !== undefined ) {
+			newData = {
+				...newData,
+				core: {
+					...newData?.core,
+					[ setting ]: data[ 'core/edit-post' ][ setting ],
+				},
+			};
+			delete newData[ 'core/edit-post' ][ setting ];
+		}
+
+		if ( data?.[ 'core/edit-site' ]?.[ setting ] !== undefined ) {
+			delete newData[ 'core/edit-site' ][ setting ];
+		}
+	} );
+
+	if ( Object.keys( newData?.[ 'core/edit-post' ] ?? {} )?.length === 0 ) {
+		delete newData[ 'core/edit-post' ];
+	}
+
+	if ( Object.keys( newData?.[ 'core/edit-site' ] ?? {} )?.length === 0 ) {
+		delete newData[ 'core/edit-site' ];
+	}
+
+	return newData;
+}