+++ packages/react-native-editor/CHANGELOG.md
@@ -17,6 +17,8 @@ For each user feature we should also add a importance categorization label  to i
 -   [**] Fix crash when sharing unsupported media types on Android [#56791]
 -   [**] Fix regressions with wrapper props and font size customization [#56985]
 -   [***] Avoid keyboard dismiss when interacting with text blocks [#57070]
+-   [**] Auto-scroll upon block insertion [#57273]
+-   [*] Unselect blocks using the hardware back button (Android) [#57279]
 
 ## 1.109.3
 -   [**] Fix duplicate/unresponsive options in font size settings. [#56985]